feat: restructure code to use Cobra for command-line interface and add version command
