+++ .gitignore
@@ -3,3 +3,4 @@
 
 picture-arrange
 dist/
+test/
