+++ README-ja.md
@@ -1,51 +1,81 @@
 # Picture Arrange
 
-このGoプログラムは、写真のEXIF日付情報を抽出し、撮影日ごとにフォルダに整理します。JPGやDNGなど、さまざまなファイル形式に対応しており、指定したディレクトリ内のすべてのファイルを処理することができます。
+このGoプログラムは、写真のEXIF日付情報を抽出し、撮影日ごとにフォルダに整理します。JPGやARWなど、さまざまなファイル形式に対応しており、指定したディレクトリ内のすべてのファイルを処理することができます。
 
 ## 特徴
 
 - 写真のEXIFデータから日付情報を抽出
 - 写真を日付ごとのフォルダに整理
-- 複数のファイル拡張子に対応（例: JPG, DNG, PNG）
+- 複数のファイル拡張子に対応（例: JPG, ARW, RAF）
 - 進捗バーで処理状況を表示
+- カメラのモデル情報に基づいた整理オプション
 
 たとえば、整理前のフォルダが以下のようになっているとします。
 
 ```
 /photos
-    ├── IMG_0001.JPG //2024-08-30
-    ├── IMG_0001.DNG //2024-08-30
-    ├── IMG_0002.JPG //2024-08-30
-    ├── IMG_0002.DNG //2024-08-30
-    ├── IMG_0003.JPG //2024-09-01
-    ├── IMG_0003.DNG //2024-09-01
-    ├── DSC_1234.DNG //2024-08-31
-    └── vacation_2024.JPG //2024-08-30
+    ├── DSC00001.JPG //2024-08-30
+    ├── DSC00001.ARW //2024-08-30
+    ├── DSC00002.JPG //2024-08-30
+    ├── DSC00002.ARW //2024-08-30
+    ├── DSC00003.JPG //2024-09-01
+    ├── DSC00003.ARW //2024-09-01
+    ├── DSC00004.ARW //2024-08-31
+    └── DSCF0590.JPG //2024-08-30
 ```
 
 これに対して、以下のようなコマンドを使うと、このように整理されます。
 
 ```bash
-./picture-arrange -dir="/photos"
+picture-arrange -d="/photos"
 ```
 
 ```
 /photos
     ├── JPG
     │   ├── 2024-08-30
-    │   │   ├── IMG_0001.JPG
-    │   │   ├── IMG_0002.JPG
-    │   │   └── vacation_2024.JPG
+    │   │   ├── DSC00001.JPG
+    │   │   ├── DSC00002.JPG
+    │   │   └── DSCF0590.JPG
     │   └── 2024-09-01
-    │       └── IMG_0003.JPG
-    └── DNG
+    │       └── DSC00003.JPG
+    └── ARW
         ├── 2024-08-30
-        │   ├── IMG_0001.DNG
-        │   └── IMG_0002.DNG
+        │   ├── DSC00001.ARW
+        │   └── DSC00002.ARW
         ├── 2024-08-31
-        │   └── DSC_1234.DNG
+        │   └── DSC00004.ARW
         └── 2024-09-01
-            └── IMG_0003.DNG
+            └── DSC00003.ARW
+```
+
+カメラモデル別に整理する場合は、以下のようなコマンドを使います。
+
+```bash
+picture-arrange -d="/photos" -m
+```
+
+```
+/photos
+    ├── ILCE-7CM2
+    │   ├── JPG
+    │   │   ├── 2024-08-30
+    │   │   │   ├── DSC00001.JPG
+    │   │   │   └── DSC00002.JPG
+    │   │   └── 2024-09-01
+    │   │       └── DSC00003.JPG
+    │   └── ARW
+    │       ├── 2024-08-30
+    │       │   ├── DSC00001.ARW
+    │       │   └── DSC00002.ARW
+    │       ├── 2024-08-31
+    │       │   └── DSC00004.ARW
+    │       └── 2024-09-01
+    │           └── DSC00003.ARW
+    └── X-T5
+        └── JPG
+            └── 2024-08-31
+                └── DSCF0590.JPG
 ```
 
 ## インストール
@@ -65,34 +95,50 @@
 
 ダウンロードしたバイナリをコマンドラインから以下のオプションを指定して実行します:
 
-- `-ext`: 処理するファイル拡張子をカンマ区切りで指定します（デフォルトは `JPG,DNG`）。
-- `-dir`: 処理するディレクトリを指定します（デフォルトはカレントディレクトリ）。
+- `-e, --extensions`: 処理するファイル拡張子をカンマ区切りで指定します（デフォルトは `JPG,ARW`）。
+- `-d, --directory`: 処理するディレクトリを指定します（デフォルトはカレントディレクトリ）。
+- `-m, --model`: カメラモデル情報を使用してファイルを整理します。
+- `version`: プログラムのバージョン情報を表示します。
 
 ### 使用例
 
 1. **デフォルトの拡張子でカレントディレクトリ内の写真を整理する:**
 
    ```bash
-   ./picture-arrange
+   picture-arrange
    ```
 
-2. **特定のディレクトリ内の写真を整理し、PNGファイルを処理する:**
+2. **特定のディレクトリ内の写真を整理し、RAFファイルを処理する:**
 
    ```bash
-   ./picture-arrange -dir="/path/to/photos" -ext="PNG"
+   picture-arrange -d="/path/to/photos" -e="RAF"
    ```
 
-3. **カレントディレクトリ内のJPGとPNGファイルを整理する:**
+3. **カレントディレクトリ内のJPGとRAFファイルを整理する:**
 
    ```bash
-   ./picture-arrange -ext="JPG,PNG"
+   picture-arrange -e="JPG,RAF"
+   ```
+
+4. **カメラモデル別に写真を整理する:**
+
+   ```bash
+   picture-arrange -m
+   ```
+
+5. **バージョン情報を表示する:**
+
+   ```bash
+   picture-arrange version
    ```
 
 ## 動作概要
 
 1. プログラムは指定したディレクトリ内で、指定された拡張子に一致するファイルをスキャンします。
 2. 各ファイルのEXIFメタデータからオリジナルの日付を抽出します。
-3. ファイルは日付ごとのフォルダ（例: `2024-08-31`）に移動されます。
+3. モードに応じてファイルを整理します:
+   - 通常モード: ファイルは拡張子別のフォルダと日付ごとのフォルダ（例: `JPG/2024-08-31`）に移動されます。
+   - モデルモード: ファイルはカメラモデル、拡張子、日付の階層（例: `ILCE-7CM2/JPG/2024-08-31`）で整理されます。
 4. 進捗バーで処理状況が表示されます。
 
 ## 依存関係
@@ -101,7 +147,8 @@
 
 - [`github.com/rwcarlsen/goexif/exif`](https://pkg.go.dev/github.com/rwcarlsen/goexif) - 画像ファイルのEXIFメタデータを読み取るためのライブラリ
 - [`github.com/schollz/progressbar/v3`](https://pkg.go.dev/github.com/schollz/progressbar/v3) - ファイル処理中に進捗バーを表示するためのライブラリ
+- [`github.com/spf13/cobra`](https://pkg.go.dev/github.com/spf13/cobra) - コマンドラインインターフェイスを構築するためのライブラリ
 
 ## ライセンス
 
-このプロジェクトはMITライセンスの下で公開されています。詳細については [LICENSE](LICENSE) ファイルをご覧ください。
+このプロジェクトはMITライセンスの下で公開されています。詳細については LICENSE ファイルをご覧ください。
